Added localized messages to CancelQuestAction.

+++ src/main/java/dev/hugog/minecraft/wonderquests/actions/implementation/CancelQuestAction.java
@@ -9,46 +9,80 @@ import dev.hugog.minecraft.wonderquests.data.keys.PlayerQuestKey;
 import dev.hugog.minecraft.wonderquests.data.services.ActiveQuestsService;
 import dev.hugog.minecraft.wonderquests.events.ActiveQuestUpdateEvent;
 import dev.hugog.minecraft.wonderquests.events.QuestUpdateType;
+import dev.hugog.minecraft.wonderquests.language.Messaging;
+import java.util.concurrent.CompletableFuture;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-public class CancelQuestAction extends AbstractAction<Boolean> {
+public class CancelQuestAction extends AbstractAction<CompletableFuture<Boolean>> {
 
   private final int questId;
 
+  private final Messaging messaging;
   private final ActiveQuestsService activeQuestsService;
   private final WonderQuests plugin;
   private final ConcurrencyHandler concurrencyHandler;
 
   @Inject
   public CancelQuestAction(@Assisted CommandSender sender, @Assisted int questId,
+      Messaging messaging,
       ActiveQuestsService activeQuestsService, WonderQuests plugin,
       ConcurrencyHandler concurrencyHandler) {
+
     super(sender);
     this.questId = questId;
+    this.messaging = messaging;
     this.activeQuestsService = activeQuestsService;
     this.plugin = plugin;
     this.concurrencyHandler = concurrencyHandler;
+
   }
 
   @Override
-  public Boolean execute() {
+  public CompletableFuture<Boolean> execute() {
 
     if (!(sender instanceof Player player)) {
-      sender.sendMessage("Only players can create quests.");
-      return false;
+      sender.sendMessage(messaging.getLocalizedChatWithPrefix("general.players_only"));
+      return CompletableFuture.completedFuture(false);
     }
 
-    activeQuestsService.removeQuest(new PlayerQuestKey(player.getUniqueId(), questId));
+    // Send pending message to provide feedback to the player
+    player.sendMessage(messaging.getLocalizedChatWithPrefix("actions.quest.abort.pending"));
 
-    // Call the event to notify that the quest has been cancelled
-    concurrencyHandler.run(() -> plugin.getServer().getPluginManager()
-        .callEvent(new ActiveQuestUpdateEvent(player, QuestUpdateType.CANCELLED)), true);
+    return activeQuestsService.hasAlreadyStartedQuest(player.getUniqueId(), questId)
+        .thenCompose(hasStarted -> {
 
-    // TODO: Add message to messages.yml
-    player.sendMessage("Quest cancelled.");
+          if (!hasStarted) {
+            player.sendMessage(messaging.getLocalizedChatWithPrefix("actions.quests.abort.not_found"));
+            return CompletableFuture.completedFuture(false);
+          }
 
-    return true;
+          return activeQuestsService.removeQuest(new PlayerQuestKey(player.getUniqueId(), questId))
+              .thenApply(ignored -> {
+
+                // Call the event to notify that the quest has been cancelled
+                concurrencyHandler.run(() -> plugin.getServer().getPluginManager()
+                        .callEvent(new ActiveQuestUpdateEvent(player, QuestUpdateType.CANCELLED)),
+                    true);
+
+                player.sendMessage(
+                    messaging.getLocalizedChatWithPrefix("actions.quests.abort.success")
+                );
+
+                return true;
+
+              })
+              .exceptionally(throwable -> {
+                player.sendMessage(messaging.getLocalizedChatWithPrefix("actions.quests.abort.error"));
+                return false;
+              });
+
+
+        })
+        .exceptionally(throwable -> {
+          player.sendMessage(messaging.getLocalizedChatWithPrefix("actions.quests.abort.error"));
+          return false;
+        });
 
   }
